generative/app/requirements.txt
pyTelegramBotAPI==4.14.1
mysql-connector-python==8.0.32
python-dotenv==1.0.1
openai==1.91.0
requests==2.31.0
certifi>=2024.2.2
charset-normalizer>=3.3.2
idna>=3.6
urllib3>=2.2.1